+++ py_experimenter/result_processor.py
@@ -6,7 +6,7 @@ from typing import List, Tuple
 import py_experimenter.utils as utils
 from py_experimenter.database_connector_lite import DatabaseConnectorLITE
 from py_experimenter.database_connector_mysql import DatabaseConnectorMYSQL
-from py_experimenter.py_experimenter_exceptions import InvalidConfigError, InvalidResultFieldError
+from py_experimenter.exceptions import InvalidConfigError, InvalidResultFieldError
 
 result_logger = logging.getLogger('result_logger')
 result_logger.setLevel(logging.INFO)
